feat(d3d11): Scale up compute shader dispatches

+++ libs/ymir-core/include/ymir/hw/vdp/renderer/vdp_renderer_d3d11.hpp
@@ -155,8 +155,8 @@ private:
     config::VDP2DebugRender &m_vdp2DebugRenderOptions;
     bool m_restoreState;
 
-    uint16 m_currScaleH = 1;
-    uint16 m_currScaleV = 1;
+    uint32 m_currScaleH = 1;
+    uint32 m_currScaleV = 1;
 
     // -------------------------------------------------------------------------
     // Configuration
